Update uploadRouter.js
Updated upload router for the automatic deletion of cloudinary assets by adding a setTimeOut() function in it which will work after set time which is 3 days (3 days -2min*,  3 days minus 2 minutes) because TTl index for the posts documents is 3 days and if the document gets deleted earlier. If we don't have the publicId we can't delete the assets from here. there are other ways to do this task like schedules (node-cron). if you set a date on the scheduler here you you can't put the year so the scheduler will do the job but it will repeat it for every year which is not required